% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_wrappers.R
\name{base_select}
\alias{base_select}
\title{Select columns from data frame}
\usage{
base_select(data, cols)
}
\arguments{
\item{data}{data.frame, tibble, or data.table.}

\item{cols}{Column names.}
}
\value{
If \code{data} is a data.table:

\code{data[, cols, with = FALSE]}

If \code{data} is a data.frame or tibble:

\code{data[, cols]}
}
\description{
Wraps \code{\link[base:Extract]{[}} for selecting columns in data frames.

Adds \code{with = FALSE} is data is a \code{data.table}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
\keyword{internal}
