% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if.R
\name{do_if}
\alias{do_if}
\alias{stop_if}
\alias{warn_if}
\alias{message_if}
\alias{identity_if}
\title{Simple do-if functions}
\usage{
stop_if(condition, message = NULL)

warn_if(condition, message = NULL)

message_if(condition, message = NULL)

identity_if(condition, x, otherwise = invisible())

do_if(condition, fn, ..., otherwise = invisible())
}
\arguments{
\item{condition}{The condition to check. (Logical)}

\item{message}{Message. (Character)

Note: If \code{NULL}, the \code{condition} will be used as message.}

\item{x}{Any object.}

\item{otherwise}{Value to return when \code{condition} is \code{FALSE}.}

\item{fn}{Any function.}

\item{...}{Arguments for \code{fn}.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

If the \code{condition} is \code{TRUE}, do something with the supplied object.
}
\details{
When \code{condition} is \code{FALSE}, they return \code{NULL} invisibly.

When \code{condition} is \code{TRUE}:

\subsection{stop_if()}{
Throws error with the supplied message.
}
\subsection{warn_if()}{
Throws warning with the supplied message.
}
\subsection{message_if()}{
Generates message with the supplied message.
}
\subsection{identity_if()}{
Returns \code{x}.
}
\subsection{do_if()}{
Calls \code{fn} with the arguments in \code{...} and returns the result.
}
}
\examples{
# Attach packages
library(rtilities2)
\donttest{
a <- 0
stop_if(a == 0, "'a' cannot be 0.")
warn_if(a == 0, "'a' was 0.")
message_if(a == 0, "'a' was so kind to be 0.")
return_if(a == 0, a)
do_if(a == 0, function(x, y) {
  x + y
}, x = 2, y = 10)
}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
