% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/not_in.R
\name{\%ni\%}
\alias{\%ni\%}
\alias{not_in}
\title{Negated value matching}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{Vector with values to be matched.}

\item{table}{Vector with values to be matched against.}
}
\value{
A logical vector, indicating if no match
was located for each element of \code{x}.
Values are \code{TRUE} or \code{FALSE} and never \code{NA}.
}
\description{
Negated \code{\link[base:match]{\%in\%}}.
}
\details{
The negation of \code{\link[base:match]{\%in\%}}.

Implementated as: \code{!(x \%in\% table)}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
