% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_between.R
\name{is_between}
\alias{is_between}
\title{Is between}
\usage{
is_between(x, lower, upper, include_limits = FALSE, allow_na = TRUE)
}
\arguments{
\item{x}{Number(s) to test. (Numeric)}

\item{lower}{Lower limit. (Numeric)}

\item{upper}{Upper limit. (Numeric)}

\item{include_limits}{Whether to include limits in
the allowed values for \code{x}. (Logical)}

\item{allow_na}{Whether to allow \code{x} to contain \code{NA}s.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

Tests if \code{x} is between \code{lower} and \code{upper}.
By default, the limits are not included.
}
\details{
When include_limits is FALSE:
\code{x > lower & x < upper}

When include_limits is TRUE:
\code{x >= lower & x <= upper}
}
\examples{
# Attach packages
library(rtilities2)

is_between(3, 1, 5) # TRUE
is_between(-1, 1, 5) # FALSE
is_between(1, 1, 5) # FALSE
is_between(1, 1, 5, include_limits = TRUE) # TRUE
is_between(5, 1, 5) # FALSE
is_between(5, 1, 5, include_limits = TRUE) # TRUE
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
