% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strip.R
\name{strip}
\alias{strip}
\title{Strips strings of non-alphanumeric characters}
\usage{
strip(strings, replacement = "", remove_spaces = FALSE, allow_na = TRUE)
}
\arguments{
\item{strings}{Vector of strings. (Character)}

\item{replacement}{What to replace blocks of punctuation with. (Character)}

\item{remove_spaces}{Whether to remove all whitespaces. (Logical)}

\item{allow_na}{Whether to allow \code{strings} to contain \code{NA}s. (Logical)}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
\enumerate{
\item Removes any character that is not alphanumeric or a space.
\item Reduces multiple consequtive whitespaces to a single whitespace.
}

Can for instance be used to simplify error messages before checking them.
}
\details{
\enumerate{
\item \code{gsub("[^[:alnum:][:blank:]]", replacement, strings))}
\item \code{gsub("[[:blank:]]+", " ", strings)} (Or \code{""} if \code{remove_spaces} is \code{TRUE})
}
}
\examples{
# Attach packages
library(rtilities2)

strip(c(
  "Hello! I am George.  \n\rDon't call me Frank!",
  "    \tAs that, is, not, my, name!"
))

strip(c(
  "Hello! I am George.  \n\rDon't call me Frank!",
  "    \tAs that, is, not, my,     name!"
), remove_spaces = TRUE)
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
