% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_if.R
\name{do_if}
\alias{do_if}
\alias{stop_if}
\alias{warn_if}
\alias{message_if}
\alias{return_if}
\title{Simple do-if functions}
\usage{
stop_if(condition, message)

warn_if(condition, message)

message_if(condition, message)

return_if(condition, x)

do_if(condition, fn, ...)
}
\arguments{
\item{condition}{The condition to check. (Logical)}

\item{message}{Message. (Character)}

\item{x}{Any object.}

\item{fn}{Any function.

E.g. \code{function(){3+5}}}

\item{...}{Arguments for \code{fn}.}
}
\description{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}

If the \code{condition} is \code{TRUE}, do something with the supplied object.
}
\details{
When \code{condition} is \code{FALSE}, they return \code{NULL} invisibly.

When \code{condition} is \code{TRUE}:

\subsection{stop_if()}{
Throw error with the supplied message.
}
\subsection{warn_if()}{
Throw warning with the supplied message.
}
\subsection{message_if()}{
Generate message with the supplied message.
}
\subsection{return_if()}{
Return \code{x}.
}
\subsection{do_if()}{
Call \code{fn} with the arguments in \code{...} and return the result.
}
}
\examples{
# Attach packages
library(rtilities2)
\donttest{
a <- 0
stop_if(a == 0, "'a' cannot be 0.")
warn_if(a == 0, "'a' was 0.")
message_if(a == 0, "'a' was so kind to be 0.")
return_if(a == 0, a)
do_if(a == 0, function(x, y){x + y}, x = 2, y = 10)
}
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}
